added hook.useUser.ts --- logic from SettingsPage.tsx; code refactoring
