+++ src/hooks/index.ts
@@ -1,1 +1,2 @@
-export * from './useTypedSelector'
+export * from './hook.useTypedSelector';
+export * from './hook.useUser';
