+++ src/pages/SettingsPage/SettingsPage.tsx
@@ -1,58 +1,25 @@
 import React, {FC, useEffect} from "react";
 import {ProgressBar} from "../../components";
 import {Button} from "@mui/material";
-import {ROUTES_PATH} from "../../constants";
-import {useNavigate} from "react-router-dom";
-import {useDispatch} from "react-redux";
-import {useTypedSelector} from "../../hooks";
-import {fetchLoggedUserInfo, resetLoginUserState} from "../../store/action-creators/user";
-import {resetPostsState} from "../../store/action-creators/post";
-import {clearLocalStorage, setLocalStorageUser} from "../../helpers/localStorageFuncs";
-
+import {useTypedSelector, useUser} from "../../hooks";
 import {LazyLoadImage} from 'react-lazy-load-image-component';
 import 'react-lazy-load-image-component/src/effects/blur.css';
 import RefreshIcon from '@mui/icons-material/Refresh';
 import LogoutIcon from '@mui/icons-material/Logout';
-import {IToken} from "../../types";
+
 
 export const SettingsPage: FC = () => {
-    const navigate = useNavigate();
-    const dispatch = useDispatch();
     const {user, isLoading} = useTypedSelector(state => state.user);
+    const {onUpdateUserDataHandler, onLogoutHandler} = useUser();
 
     const renderCondition = user !== null;
 
     useEffect(() => {
         // load data if extra user info is absent
-        if (!user?.crypto?.coin) {
+        if (!isLoading && !user?.crypto?.coin) {
             onUpdateUserDataHandler();
         }
     }, [])
-
-    const onLogoutHandler = () => {
-        clearLocalStorage();
-        dispatch(resetLoginUserState());
-        dispatch(resetPostsState());
-        navigate(`/${ROUTES_PATH.login}`);
-    };
-
-    const updateTokenAndReFetchCallBack = (token: IToken) => {
-        setLocalStorageUser({
-            ...user,
-            ...(token ?? {}),
-        });
-
-        dispatch(fetchLoggedUserInfo(token.accessToken, token.refreshToken));
-    };
-
-    const onUpdateUserDataHandler = () => {
-        dispatch(fetchLoggedUserInfo(
-            user.accessToken,
-            user.refreshToken,
-            updateTokenAndReFetchCallBack,
-            onLogoutHandler
-        ));
-    };
 
     const renderContent = () => {
         return (
